pyproject.toml
[tool.poetry]
name = "mapbot-py"
version = "0.1.0"
description = ""
authors = ["TetraspaceW <holomanga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
geopy = "^2.4.1"
supabase = "^2.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
